requirements.txt
fitz==0.0.1.dev2
Flask==3.0.3
Pillow==10.3.0
pytesseract==0.3.10
PyMuPDF==1.24.3
opencv-python==4.9.0.80

